Camera and board setup
There now exisits a Camera which is controlled by the left and right
keys, and there is a model of the board being displayed in the center of
the screen. If there appear to be missing files right-click in the
Solution explorer on Lotus and select add exisiting item and then add
the .cs files. Perform a similar process for the images and font only
right-click on LotusContent add a folder and then add exisiting items to
that folder.