Mouse Location tracking
The mouse coordinates on the (y=0) plane are now being displayed and can
be used to locate positions on the board. The top of the board is the
(y=0) plane just FYI. Also created a graph classes that will be used to
map out all positions that pieces can land on. i've implement a small
feature that displays which node on the graph is under the mouse.