Chat Program & Networking Library
The chat program is untested. In order to use the Lidgren networking
library, go File->Add->Existing Project and select Lidgren.Network ->
Lidgren.Network.csproj and add to the references ("projects" tab) of the
project you want to add it to (once it's finished, the chat program, to
check it out; once we have networking implemented/started, the game
itself).